Cut search space by using a rough energy calculation
energy is the remaining flow in the system
score is the score so far in the current search branch
minute is the number of minutes left
max_score is the maximum score we have found so far

If at any point the following inequality holds

score + energy * minute < max_score

Then we know this branch can't beat the highest score and can abort
the branch.

It's a very rough calculation, but does restrict the search space
enough for it to solve the problem in a few minutes.