Solve 2023 Day 25
Looked up algorithms for finding a minimum cut in a graph
Karger's algorithm is clearly the right one for this problem
as we know the cut size and can just keep running it until
we find the right cut.

However, I've never seen Stoer-Wagner before, so thought I'd
try that. This is a bad implementation that doesn't use the
priority queue (or exit early once we find a cut of size 3)